Move Manipulator class
